First Commit
The project have the ability to get student data and store it in a vector and each student object can store courses objects using a vector. The manager class the methods to add students to static vector, add courses to the any student in the students student, and display details of every student along with their registered courses.