Added the functions Manager and DB_Manager classes
For Manager class:\nThe function to add student to students vector is added, added the function to add course to courses vector, and added the function to register student to a course. The previous function in DB_Manager were changed to support these updated functions in Manager class and added addCourseToStudent function in DB_Manager to register student to course in db.